api-ferreterias/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
anthropic>=0.34.0
openai>=1.30.0
requests>=2.31.0
httpx>=0.27.0
openpyxl==3.1.5
rapidfuzz>=3.9.0
